+++ app/javascript/features/project/chart/project-stats.module.scss
