+++ app/javascript/features/project/form/new-project.module.scss
